+++ scripts/publish-all.sh
@@ -5,4 +5,4 @@ SCRIPT_DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )"
 cd $SCRIPT_DIR/..
 
 cargo publish -p insta
-cd cargo-insta && cargo publish
+cargo publish -p cargo-insta